+++ src/grammar.h
@@ -1,0 +1,11 @@
+#include <string>
+
+namespace score { namespace ast {
+  struct Statement { };
+}
+}
+
+namespace score
+{
+  bool parse(ast::Statement& out, const std::string& in);
+}
